Updated list of stops used
There were some new stops built that the hardcoded list didn't include. The new hardcoded list is exhaustive and I've added an API call (that is unused) that can be used to get the latest. There's no need to grab the list every time because it changes so infrequently.

Also cleaned up the gradle files, removed all the Kotlin references